requirements.txt
kafka-python==2.0.2
pandas==1.3.3
pyspark==3.2.1
pytest==7.1.1
python-dotenv==0.20.0
